report number missed frame

+++ src/AOloopControl_acquireCalib/AOloopControl_acquireCalib.c
@@ -1376,10 +1376,10 @@ long AOloopControl_acquireCalib_Measure_WFSrespC(
                 {
                     printf("Warning: missed %d frame(s)\n", NBmissedframe);
                     printf("     PokeIndex   %d / %d\n", (int) PokeIndex, (int) NBpoke);
-                    printf("     kk          %d\n", kk);
-                    printf("     NBave       %d\n", NBave);
-                    printf("     kk1         %d\n", kk1);
-                    printf("     PokeIndex1  %d\n", PokeIndex1);
+                    printf("     kk          %d\n", (int) kk);
+                    printf("     NBave       %d\n", (int) NBave);
+                    printf("     kk1         %d\n", (int) kk1);
+                    printf("     PokeIndex1  %d\n", (int) PokeIndex1);
                 }
